DefaultDescriptorValidationRule
